move query constructors to a separate package
