+++ moonlight-core/src/main/scala/com/github/jamedge/moonlight/core/model/neo4j/queryconstructors/LineQueriesConstructor.scala
@@ -1,6 +1,8 @@
-package com.github.jamedge.moonlight.core.model.neo4j
+package com.github.jamedge.moonlight.core.model.neo4j.queryconstructors
 
+import com.github.jamedge.moonlight.core.model.neo4j.Nodes
 import neotypes.DeferredQueryBuilder
+
 import neotypes.implicits.all._
 
 object LineQueriesConstructor
