+++ moonlight-core/src/main/scala/com/github/jamedge/moonlight/core/model/neo4j/queryconstructors/LineageQueriesConstructor.scala
@@ -1,6 +1,7 @@
-package com.github.jamedge.moonlight.core.model.neo4j
+package com.github.jamedge.moonlight.core.model.neo4j.queryconstructors
 
 import neotypes.DeferredQueryBuilder
+
 import neotypes.implicits.all._
 
 object LineageQueriesConstructor {
