+++ moonlight-core/src/main/scala/com/github/jamedge/moonlight/core/service/lineage/LineagePersistenceLayer.scala
@@ -1,7 +1,7 @@
 package com.github.jamedge.moonlight.core.service.lineage
 
+import com.github.jamedge.moonlight.core.model.neo4j.queryconstructors.LineageQueriesConstructor
 import com.github.jamedge.moonlight.core.model.{IOElement, Storage}
-import com.github.jamedge.moonlight.core.model.neo4j.LineageQueriesConstructor
 import neotypes.Driver
 import org.neo4j.driver.v1.Value
 import scalax.collection.Graph
